+++ projects/libraries/streamix/angular/src/lib/index.ts
@@ -1,7 +1,6 @@
 export * from './compiled-block-helpers';
 export * from './compiled-block';
 export * from './structural-block';
-export * from './static-block';
 export * from './compiled-view';
 export * from './binding-table';
 export * from './direct-binding';